Consolidate batch recorder tests and fixes bug
